# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Retro Gamer Blog API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  OWNER: 'OWNER',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type RoleUpdateRole = typeof RoleUpdateRole[keyof typeof RoleUpdateRole];


export const RoleUpdateRole = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  OWNER: 'OWNER',
} as const;

export interface RoleUpdate {
  role: RoleUpdateRole;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  postCount?: number;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  content: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  authorId: number;
  author: User;
  /** @nullable */
  categoryId?: number | null;
  category?: Category | null;
  tags?: string[];
  isPinned: boolean;
  isFeatured: boolean;
  viewCount: number;
  likeCount: number;
  commentCount: number;
  likedByMe?: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PostsPage {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}

export interface PostInput {
  /** @minLength 1 */
  title: string;
  content: string;
  excerpt?: string;
  coverImageUrl?: string;
  categoryId?: number;
  tags?: string[];
  isPinned?: boolean;
  isFeatured?: boolean;
}

export interface PostUpdate {
  title?: string;
  content?: string;
  excerpt?: string;
  coverImageUrl?: string;
  /** @nullable */
  categoryId?: number | null;
  tags?: string[];
  isPinned?: boolean;
  isFeatured?: boolean;
}

export interface LikeResult {
  liked: boolean;
  likeCount: number;
}

export interface Comment {
  id: number;
  content: string;
  postId: number;
  authorId: number;
  author: User;
  createdAt: string;
}

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}

export interface CategoryInput {
  name: string;
  description?: string;
  color?: string;
}

export interface CategoryUpdate {
  name?: string;
  description?: string;
  color?: string;
}

export interface Tag {
  name: string;
  postCount: number;
}

export interface AnalyticsOverview {
  totalPosts: number;
  totalUsers: number;
  totalComments: number;
  totalViews: number;
  totalLikes: number;
  postsThisWeek: number;
  commentsThisWeek: number;
  topPosts: Post[];
}

export interface OnlineCount {
  count: number;
}

export interface Notification {
  id: number;
  type: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  postId?: number | null;
  createdAt: string;
}

export type GetPostsParams = {
page?: number;
limit?: number;
categoryId?: number;
tag?: string;
search?: string;
pinned?: boolean;
};