# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Retro Gamer Blog API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: Date;
}